+++ bundlewatch.config.js
@@ -11,11 +11,10 @@ const packageFolders = [
 ];
 
 const excludedPackages = [
-    'packages/services/f-metadata',
+    'packages/services/f-braze-adapter',
     'packages/tools/generator-component',
     'packages/tools/storybook',
-    'packages/services/f-wdio-utils',
-    'packages/components/organisms/f-status-banner' // Excluding as the footer is pushing this package over the 100kb.
+    'packages/services/f-wdio-utils'
 ];
 
 /**