+++ packages/components/organisms/f-checkout/src/components/_tests/Checkout.test.js
@@ -194,7 +194,8 @@ describe('Checkout', () => {
                 expect(errorPage.exists()).toBe(false);
             });
 
-            it('should render the error page and not the checkout form when set to `false`', () => {
+
+            it('should render the error page and not the checkout form and spinner when set to `false`', () => {
                 // Arrange & Act
                 const wrapper = mount(VueCheckout, {
                     i18n,
@@ -203,17 +204,45 @@ describe('Checkout', () => {
                     propsData,
                     data () {
                         return {
-                            hasCheckoutLoadedSuccessfully: false
+                            hasCheckoutLoadedSuccessfully: false,
+                            isLoading: false
                         };
                     }
                 });
 
                 const checkoutForm = wrapper.find('[data-test-id="checkout-component"]');
                 const errorPage = wrapper.find('[data-test-id="checkout-error-page-component"]');
+                const spinner = wrapper.find('[data-test-id="checkout-loading-spinner"]');
 
                 // Assert
                 expect(checkoutForm.exists()).toBe(false);
+                expect(spinner.exists()).toBe(false);
                 expect(errorPage.exists()).toBe(true);
+            });
+        });
+
+        describe('shouldShowSpinner ::', () => {
+            it('should render the loading spinner and not the checkout form when shouldShowSpinner is set to `true`', () => {
+                // Arrange & Act
+                const wrapper = mount(VueCheckout, {
+                    i18n,
+                    store: createStore(),
+                    localVue,
+                    propsData,
+                    data () {
+                        return {
+                            hasCheckoutLoadedSuccessfully: true,
+                            shouldShowSpinner: true
+                        };
+                    }
+                });
+
+                const checkoutForm = wrapper.find('[data-test-id="checkout-component"]');
+                const spinner = wrapper.find('[data-test-id="checkout-loading-spinner"]');
+
+                // Assert
+                expect(checkoutForm.exists()).toBe(false);
+                expect(spinner.exists()).toBe(true);
             });
         });
     });
@@ -499,6 +528,40 @@ describe('Checkout', () => {
                 expect(loadAvailableFulfilmentSpy).toHaveBeenCalled();
             });
 
+            it('should call `startSpinnerCountdown`', async () => {
+                // Arrange
+                const startSpinnerCountdownSpy = jest.spyOn(VueCheckout.methods, 'startSpinnerCountdown');
+
+                // Act
+                shallowMount(VueCheckout, {
+                    store: createStore(),
+                    i18n,
+                    localVue,
+                    propsData
+                });
+                await flushPromises();
+
+                // Assert
+                expect(startSpinnerCountdownSpy).toHaveBeenCalled();
+            });
+
+            it('should call `resetLoadingState`', async () => {
+                // Arrange
+                const resetLoadingStateSpy = jest.spyOn(VueCheckout.methods, 'resetLoadingState');
+
+                // Act
+                shallowMount(VueCheckout, {
+                    store: createStore(),
+                    i18n,
+                    localVue,
+                    propsData
+                });
+                await flushPromises();
+
+                // Assert
+                expect(resetLoadingStateSpy).toHaveBeenCalled();
+            });
+
             describe('if shouldLoadAddress returns `false`', () => {
                 it('should not call `loadAddress`', async () => {
                     // Arrange & Act
@@ -570,6 +633,106 @@ describe('Checkout', () => {
 
                     // Assert
                     expect(loadCheckoutSpy).toHaveBeenCalled();
+                });
+            });
+        });
+
+        describe('startSpinnerCountdown ::', () => {
+            beforeEach(() => {
+                jest.spyOn(VueCheckout.methods, 'initialise').mockImplementation();
+                jest.useFakeTimers();
+            });
+
+            afterEach(() => {
+                jest.clearAllMocks();
+                jest.clearAllTimers();
+            });
+
+            describe('when `isLoading` is `true`', () => {
+                it('should not set `shouldShowSpinner` to `true` before one second', () => {
+                    // Arrange & Act
+                    const wrapper = shallowMount(VueCheckout, {
+                        store: createStore(),
+                        i18n,
+                        localVue,
+                        propsData,
+                        data () {
+                            return {
+                                isLoading: true
+                            };
+                        }
+                    });
+
+                    wrapper.vm.startSpinnerCountdown();
+                    jest.advanceTimersByTime(999);
+
+                    // Assert
+                    expect(wrapper.vm.shouldShowSpinner).toBe(false);
+                });
+
+                it('should set `shouldShowSpinner` to `true` after one second', () => {
+                    // Arrange & Act
+                    const wrapper = shallowMount(VueCheckout, {
+                        store: createStore(),
+                        i18n,
+                        localVue,
+                        propsData,
+                        data () {
+                            return {
+                                isLoading: true
+                            };
+                        }
+                    });
+
+                    wrapper.vm.startSpinnerCountdown();
+                    jest.advanceTimersByTime(1000);
+
+                    // Assert
+                    expect(wrapper.vm.shouldShowSpinner).toBe(true);
+                });
+            });
+
+            describe('when `isLoading` is `false`', () => {
+                it('should not set `shouldShowSpinner` to `true` after one second', () => {
+                    // Arrange & Act
+                    const wrapper = shallowMount(VueCheckout, {
+                        store: createStore(),
+                        i18n,
+                        localVue,
+                        propsData,
+                        data () {
+                            return {
+                                isLoading: false
+                            };
+                        }
+                    });
+
+                    wrapper.vm.startSpinnerCountdown();
+                    jest.advanceTimersByTime(1000);
+
+                    // Assert
+                    expect(wrapper.vm.shouldShowSpinner).toBe(false);
+                });
+
+                it('should not set `shouldShowSpinner` to `true` before one second', () => {
+                    // Arrange & Act
+                    const wrapper = shallowMount(VueCheckout, {
+                        store: createStore(),
+                        i18n,
+                        localVue,
+                        propsData,
+                        data () {
+                            return {
+                                isLoading: false
+                            };
+                        }
+                    });
+
+                    wrapper.vm.startSpinnerCountdown();
+                    jest.advanceTimersByTime(999);
+
+                    // Assert
+                    expect(wrapper.vm.shouldShowSpinner).toBe(false);
                 });
             });
         });