+++ packages/components/organisms/f-status-banner/CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v1.1.0
+------------------------------
+*February 23, 2021*
+
+### Removed
+- `header` & `footer` components so the consuming application can render their own.
+- Tests that covered header & footer.
+- Storybook dispatch error.
+
+
 v1.0.0
 ------------------------------
 *February 12, 2021*